requirements.txt
ag-ui-protocol==0.1.8
agno==2.0.8
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.5
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
crashtest==0.4.1
ddgs==9.5.5
distlib==0.3.9
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dulwich==0.22.8
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
fsspec==2025.9.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.38.0
googlesearch-python==1.3.0
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.10.0
keyring==25.6.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
numpy==2.3.3
openai==1.106.1
packaging==24.2
pbs-installer==2025.2.12
pdf-orientation-corrector==0.1.4
pdf2image==1.17.0
pgvector==0.4.1
pillow==10.4.0
pkginfo==1.12.1.2
platformdirs==4.3.6
pluggy==1.6.0
poetry==2.1.1
poetry-core==2.1.1
primp==0.15.0
psycopg-binary==3.2.10
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.0.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.37.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.47.3
tabulate==0.9.0
tenacity==9.1.2
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.3.3.18
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.29.2
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.23.0
